% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratewidget.R
\name{ratewidget}
\alias{ratewidget}
\title{Show the summary proportion in a widget}
\usage{
ratewidget(
  data,
  numerator,
  denominator,
  multiplier = 100,
  digits = 0,
  suffix = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{A data frame or data table containing the data to summarize,
with at least two numeric columns: one for the numerator and one for the
denominator. Normally, this should be an instance of \link[crosstalk:SharedData]{crosstalk::SharedData}.}

\item{numerator}{The name of the column in \code{data} to use as the numerator.}

\item{denominator}{The name of the column in \code{data} to use as the
denominator.}

\item{multiplier}{A numeric value to multiply the proportion by (default is
100, to express as a percentage).}

\item{digits}{The number of decimal places to round the result to (default is
0).}

\item{suffix}{An optional suffix to append to the displayed value (e.g., "\%").}

\item{width}{The width of the widget (optional).}

\item{height}{The height of the widget (optional).}

\item{elementId}{An optional element ID for the widget.}
}
\description{
A \code{ratewidget} displays a single statistic derived from a linked table, where
the numerator and denominator columns are specific. This is best used with
the \code{crosstalk}package, to allow dynamic filtering of the underlying data.
}
